api/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
billiard==3.5.0.3
boto3==1.5.25
botocore==1.8.39          # via boto3, s3transfer
certifi==2018.1.18        # via requests
chardet==3.0.4            # via requests
coreapi==2.3.3
coreschema==0.0.4         # via coreapi
coverage==4.5
daiquiri==1.3.0
django==2.0.2
djangorestframework==3.7.7
docutils==0.14            # via botocore
idna==2.6                 # via requests
itypes==1.1.0             # via coreapi
jinja2==2.10              # via coreschema
jmespath==0.9.3           # via boto3, botocore
markupsafe==1.0           # via jinja2
psycopg2==2.7.4
python-dateutil==2.6.1    # via botocore
python-nomad==0.6.1
pytz==2018.3              # via django
requests==2.18.4
retrying==1.3.3
s3transfer==0.1.12        # via boto3
six==1.11.0               # via python-dateutil, retrying
uritemplate==3.0.0        # via coreapi
urllib3==1.22             # via requests
